requirements-dev.txt
coverage
pyhamcrest
pytest>=2.8
pytest-cov
pytest-sugar
